refactor: Extract valueFormatter to use it in the story
