+++ src/domain/data.ts
@@ -292,6 +292,4 @@ export type WikiPageSlug = (typeof wikiPageSlugs)[
   | "available"
   | "missing"][number];
 
-export type ValueFormatter = (
-  value: number | boolean | string | { valueOf(): string }
-) => string;
+export type ValueFormatter = (value: number) => string;
