+++ src/pages/map/index.tsx
@@ -39,11 +39,9 @@ import {
   NetworkLevel,
   TariffCategory,
   useColorScale,
-  ValueFormatter,
 } from "src/domain/data";
-import { useFormatCurrency } from "src/domain/helpers";
+import { useIndicatorValueFormatter } from "src/domain/helpers";
 import {
-  QueryStateSunshineIndicator,
   useQueryStateEnergyPricesMap,
   useQueryStateMapCommon,
   useQueryStateSunshineMap,
@@ -79,19 +77,6 @@ export const getServerSideProps: GetServerSideProps<
   { locale: string }
 > = async ({ locale }) => {
   return { props: { locale: locale ?? defaultLocale } };
-};
-
-const indicatorFormatterType: Record<
-  QueryStateSunshineIndicator,
-  "number" | "boolean"
-> = {
-  saidi: "number",
-  saifi: "number",
-  serviceQuality: "number",
-  compliance: "boolean",
-  energyTariffs: "number",
-  netTariffs: "number",
-  networkCosts: "number",
 };
 
 const IndexPageContent = ({
@@ -230,15 +215,7 @@ const IndexPageContent = ({
     }
   }, [activeId, isSunshine]);
 
-  const formatNumber = useFormatCurrency();
-  const formatBooleanNumber = ((value: number) => {
-    return value === 1 ? t`Yes` : value === 0 ? t`No` : "-";
-  }) as ValueFormatter;
-  const formatterTypes = {
-    number: formatNumber,
-    boolean: formatBooleanNumber,
-  };
-  const valueFormatter = formatterTypes[indicatorFormatterType[indicator]];
+  const valueFormatter = useIndicatorValueFormatter(indicator);
 
   const map = isElectricityTab ? (
     <EnergyPricesMap
